+++ autonomous-scheduler/vGPUScheduler/go.mod
@@ -5,11 +5,9 @@ go 1.17
 require (
 	k8s.io/api v0.22.6
 	k8s.io/apimachinery v0.22.6
-	k8s.io/apiserver v0.22.6
 	k8s.io/client-go v0.22.6
 	k8s.io/component-base v0.22.6
 	k8s.io/klog/v2 v2.9.0
-	k8s.io/kube-scheduler v0.22.6
 	k8s.io/kubernetes v1.22.6
 	sigs.k8s.io/scheduler-plugins v0.22.6
 )
@@ -98,10 +96,12 @@ require (
 	gopkg.in/natefinch/lumberjack.v2 v2.0.0 // indirect
 	gopkg.in/yaml.v2 v2.4.0 // indirect
 	gopkg.in/yaml.v3 v3.0.0-20210107192922-496545a6307b // indirect
+	k8s.io/apiserver v0.22.6 // indirect
 	k8s.io/cloud-provider v0.22.6 // indirect
 	k8s.io/component-helpers v0.22.6 // indirect
 	k8s.io/csi-translation-lib v0.22.6 // indirect
 	k8s.io/kube-openapi v0.0.0-20211109043538-20434351676c // indirect
+	k8s.io/kube-scheduler v0.22.6 // indirect
 	k8s.io/mount-utils v0.22.6 // indirect
 	k8s.io/utils v0.0.0-20210819203725-bdf08cb9a70a // indirect
 	sigs.k8s.io/apiserver-network-proxy/konnectivity-client v0.0.27 // indirect
